[std/node] adds support for nodejs encoding options and 6 characters suffix

+++ std/node/_fs/_fs_mkdtemp_test.ts
@@ -5,34 +5,16 @@ import { existsSync } from "./_fs_exists.ts";
 import { env } from "../process.ts";
 import { isWindows } from "../../_util/os.ts";
 
-const prefix = isWindows ? env.TEMP : env.TMPDIR;
+const prefix = isWindows ? env.TEMP+'\\' : env.TMPDIR+'/';
 const doesNotExists = "/does/not/exists/";
-const options = { encoding: "utf8" };
+const options = { encoding: "ascii" };
+const badOptions = { encoding: "bogus" };
 
 Deno.test({
   name: "[node/fs] mkdtemp",
   fn: async () => {
     const result = await new Promise((resolve) => {
       mkdtemp(prefix, (err, directory) => {
-        if (err) {
-          resolve(false);
-        } else if (directory) {
-          resolve(existsSync(directory));
-          Deno.removeSync(directory);
-        } else {
-          resolve(false);
-        }
-      });
-    });
-    assert(result);
-  },
-});
-
-Deno.test({
-  name: "[node/fs] mkdtemp (with options)",
-  fn: async () => {
-    const result = await new Promise((resolve) => {
-      mkdtemp(prefix, options, (err, directory) => {
         if (err) {
           resolve(false);
         } else if (directory) {
@@ -64,19 +46,45 @@ Deno.test({
 });
 
 Deno.test({
-  name: "[node/fs] mkdtempSync",
-  fn: () => {
-    const directory = mkdtempSync(prefix);
-    const dirExists = existsSync(directory);
-    Deno.removeSync(directory);
-    assert(dirExists);
+  name: "[node/fs] mkdtemp (with options)",
+  fn: async () => {
+    const result = await new Promise((resolve) => {
+      mkdtemp(prefix, options, (err, directory) => {
+        if (err) {
+          resolve(false);
+        } else if (directory) {
+          resolve(existsSync(directory));
+          Deno.removeSync(directory);
+        } else {
+          resolve(false);
+        }
+      });
+    });
+    assert(result);
   },
 });
 
 Deno.test({
-  name: "[node/fs] mkdtempSync (with options)",
+  name: "[node/fs] mkdtemp (with bad options)",
+  fn: async () => {
+    const result = await new Promise((resolve) => {
+      try {
+        mkdtemp(prefix, badOptions, (err, directory) => {
+          // should have thrown already...
+          resolve(false);
+        });
+      } catch (error) {
+        resolve(true);
+      }
+    });
+    assert(result);
+  },
+});
+
+Deno.test({
+  name: "[node/fs] mkdtempSync",
   fn: () => {
-    const directory = mkdtempSync(prefix, options);
+    const directory = mkdtempSync(prefix);
     const dirExists = existsSync(directory);
     Deno.removeSync(directory);
     assert(dirExists);
@@ -98,3 +106,29 @@ Deno.test({
     }
   },
 });
+
+Deno.test({
+  name: "[node/fs] mkdtempSync (with options)",
+  fn: () => {
+    const directory = mkdtempSync(prefix, options);
+    const dirExists = existsSync(directory);
+    Deno.removeSync(directory);
+    assert(dirExists);
+  },
+});
+
+Deno.test({
+  name: "[node/fs] mkdtempSync (with bad options)",
+  fn: () => {
+    try {
+      const directory = mkdtempSync(prefix, badOptions);
+      // should have thrown already...
+
+      Deno.removeSync(directory);
+      assert(false);
+    } catch (error) {
+      assert(true);
+    }
+  },
+});
+
